add bigram to doc2vec
